requirements.txt
# This file is used to install the required packages for the project. Feel free to modify it as needed. You are free to add or remove packages.
# The packages listed here will be installed using pip when you run the command `pip install -r requirements.txt`
unstructured[pdf,ppt,pttx]
langchain
chromadb
requests
llama-cloud-services 
llama-index-core 
llama-index-readers-file 
python-dotenv